mediacat/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "mediacat"
version = "0.1.0"
description = "Cataloging platform for physical music media (vinyl + CD)."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "MediaCat Contributors" }]
classifiers = [
    "Private :: Do Not Upload",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
]
# Runtime dependencies are intentionally minimal at Section 1.
# Application dependencies are added in Sections 3+ as the modules land.
dependencies = [
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",
    "pyyaml>=6.0.1",
    "structlog>=24.1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "greenlet>=3.0",
    "minio>=7.2",
    "Pillow>=10.4",
    "httpx>=0.27",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "argon2-cffi>=23.1",
    "itsdangerous>=2.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "bandit>=1.7",
    "pip-audit>=2.7",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pre-commit>=3.7",
    "types-PyYAML",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-mermaid2-plugin>=1.1",
    "pdoc>=14.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mediacat"]

# ---------- Ruff ------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]
extend-exclude = ["alembic/versions", "site", "build", "dist"]

[tool.ruff.lint]
select = [
    "E", "F", "W",      # pyflakes / pycodestyle
    "I",                # isort
    "B",                # flake8-bugbear
    "C4",               # comprehensions
    "UP",               # pyupgrade
    "SIM",              # simplify
    "S",                # bandit-lite
    "N",                # pep8-naming
    "RUF",
    "PL",               # pylint
    "PT",               # pytest
    "ASYNC",
    "TID",              # tidy imports
    "ARG",              # unused args
    "DTZ",              # datetime timezones
    "LOG",              # logging
    "G",                # logging format
    "A",                # shadowing builtins
]
ignore = [
    "E501",   # line length handled by formatter
    "PLR0913",
    "PLR2004",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "PLR2004", "ARG001", "ARG002", "PLC0415"]
"alembic/env.py" = ["E402"]
"src/mediacat/rules/local.py" = ["ARG002"]  # media_format is interface-mandated
"src/mediacat/vision/adapter.py" = ["ARG002", "PLC0415"]  # interface params + deferred imports
"src/mediacat/web/routes.py" = ["ARG001"]  # FastAPI handler params used for DI/form binding
"src/mediacat/web/middleware.py" = ["PLC0415", "SIM110"]  # inline import + loop clarity
"src/mediacat/config.py" = ["PLC0415", "S104"]  # deferred imports + 0.0.0.0 bind is intentional
"src/mediacat/db/audit.py" = ["PLC0415"]  # deferred uuid import
"src/mediacat/worker.py" = ["ARG001"]  # signal handler frame param
"src/mediacat/rules/engine.py" = ["PLC0415"]  # deferred imports in factory

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------- Mypy ------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_reexport = true
show_error_codes = true
show_column_numbers = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/versions/", "build/", "dist/", "site/"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# ---------- Pytest ----------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-config",
    "--strict-markers",
    "--cov=mediacat",
    "--cov-report=term-missing",
    "--cov-report=xml",
]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: integration tests requiring docker stack",
    "slow: tests that take >1s",
]
filterwarnings = ["error"]

# ---------- Coverage --------------------------------------------------------
[tool.coverage.run]
branch = true
source = ["src/mediacat"]
omit = ["src/mediacat/__main__.py"]

[tool.coverage.report]
skip_covered = false
show_missing = true
fail_under = 55       # ratcheted upward as integration tests land
exclude_also = [
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "pragma: no cover",
]

# ---------- Bandit ----------------------------------------------------------
[tool.bandit]
exclude_dirs = ["tests", "alembic/versions", ".venv", "build", "dist", "site"]
skips = ["B101"]  # assert_used in tests handled by per-file ignore already
